Implement the new code related to D4.8.
The new objective MinCost() minimises the value of a cost functions
associated to every VM -> Node placement. It tends to minimise the
hosting costs and the use of long distance clouds.

The customisation is done through the ModelView "CostView" which states
which nodes are in the edge, which nodes represent public clouds and
the cost associated to every VM->Node.

CMinCostTest illustrates the behavior using a made up use cases
involving the initial deployment that requires to offload a bit on
public clouds and a scale down that enables to bring back the VMs in the
edge.